+++ src/common/lib/client/channelstatechange.ts
@@ -5,11 +5,21 @@ class ChannelStateChange {
   current: string;
   resumed?: boolean;
   reason?: string | Error | ErrorInfo;
+  hasBacklog?: boolean;
 
-  constructor(previous: string, current: string, resumed?: boolean, reason?: string | Error | ErrorInfo | null) {
+  constructor(
+    previous: string,
+    current: string,
+    resumed?: boolean,
+    hasBacklog?: boolean,
+    reason?: string | Error | ErrorInfo | null
+  ) {
     this.previous = previous;
     this.current = current;
-    if (current === 'attached') this.resumed = resumed;
+    if (current === 'attached') {
+      this.resumed = resumed;
+      this.hasBacklog = hasBacklog;
+    }
     if (reason) this.reason = reason;
   }
 }