+++ test/support/browser_file_list.js
@@ -64,4 +64,6 @@ window.__testFiles__.files = {
   'test/browser/connection.test.js': true,
   'test/browser/simple.test.js': true,
   'test/browser/http.test.js': true,
+  'test/rest/status.test.js': true,
+  'test/rest/batch.test.js': true,
 };