+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This contains only the most important and/or user-facing changes; for a full changelog, see the commit history.
 
+## [1.2.42](https://github.com/ably/ably-js/tree/1.2.42) (2023-07-24)
+
+- Auth: remain connected upon failed authorize unless returning explicit 403 [\#1385](https://github.com/ably/ably-js/pull/1385)
+- Make `Utils#inspectError` use `toString` for `Error`-like values [\#1391](https://github.com/ably/ably-js/pull/1391)
+- docs: fix description of AuthOptions.token [\#1368](https://github.com/ably/ably-js/pull/1368)
+
+## [1.2.41](https://github.com/ably/ably-js/tree/1.2.41) (2023-06-29)
+
+- add `ChannelStateChange.hasBacklog` and return state change to attach promise/callback [\#1347](https://github.com/ably/ably-js/pull/1347)
+- fix a bug where host fallback was initially skipped after falling back to the base transport [\#1357](https://github.com/ably/ably-js/pull/1357)
+
 ## [1.2.40](https://github.com/ably/ably-js/tree/1.2.40) (2023-05-26)
 
 This release adds a new experimental `channels.getDerived` method which allows you to create custom realtime data feeds by selectively subscribing to receive only part of the data from the channel. See the [announcement post](https://pages.ably.com/subscription-filters-preview) for more information.